-- refactoring: splitted by modules -- added posthoc: Kruskal-Wallis Dunn's multiple comparisons
