+++ src/AutoStatLib/_version.py
@@ -1,2 +1,2 @@
 # AutoStatLib package version:
-__version__ = "0.2.2"
+__version__ = "0.2.3"
